Created method for viewing pictures in large mode
